requirements.txt
streamlit
joblib
numpy
pandas
beautifulsoup4
nltk
scikit-learn
google-cloud-translate